requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
apscheduler==3.10.4
gunicorn==21.2.0
python-dotenv==1.2.1
